Regroup timezone utils in a single module
